Fixed issue where attempting to delete partial entries when it was not enabled caused errors.

+++ gp-easy-passthrough/gpep-edit-entry.php
@@ -27,7 +27,7 @@ class GPEP_Edit_Entry {
 
 		$update_entry_id = $this->get_update_entry_id( $form['id'] );
 		if ( $update_entry_id ) {
-			if ( $this->delete_partial ) {
+			if ( $this->delete_partial && $entry_id !== null ) {
 				GFAPI::delete_entry( $entry_id );
 			}
 			return $update_entry_id;
